+++ api_yamdb/api/admin.py
@@ -1,21 +1,5 @@
 from django.contrib import admin
-
-from reviews.models import User, Genre, Category, Title
-
-
-@admin.register(User)
-class UserAdmin(admin.ModelAdmin):
-    list_display = (
-        'first_name',
-        'last_name',
-        'username',
-        'role',
-        'email',
-        'bio'
-    )
-    list_editable = ('role',)
-    search_fields = ('username',)
-    empty_value_display = '-пусто-'
+from reviews.models import Category, Genre, Title
 
 
 @admin.register(Category)
