+++ api_yamdb/api/serializers.py
@@ -3,9 +3,10 @@ from django.conf import settings
 from django.core.validators import RegexValidator
 from django.shortcuts import get_object_or_404
 from rest_framework import serializers
-from reviews.models import Category, Comments, Genre, Review, Title, User
 
-from .utils import generate_short_hash_mm3
+from .utils import generate_short_hash_mm3   # isort: skip
+from reviews.models import Category, Comments, Genre, Review, Title  # isort: skip
+from users.models import User  # isort: skip
 
 
 class CategorySerializer(serializers.ModelSerializer):
