+++ api_yamdb/reviews/models.py
@@ -1,30 +1,8 @@
-from django.contrib.auth.models import AbstractUser
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
-from django.utils import timezone
 
-from .validators import year_validator
-
-ROLES = [
-    ('user', 'Пользователь'),
-    ('admin', 'Администратор'),
-    ('moderator', 'Модератор'),
-]
-
-
-class User(AbstractUser):
-    """Кастомная модель пользователя."""
-    username = models.CharField(max_length=150, unique=True)
-    first_name = models.CharField(max_length=150, blank=True)
-    last_name = models.CharField(max_length=150, blank=True)
-    email = models.EmailField(max_length=254, unique=True)
-    role = models.CharField(max_length=15,
-                            choices=ROLES,
-                            default='user',
-                            verbose_name='Роль пользователя',
-                            )
-    bio = models.TextField(verbose_name='Биография', blank=True)
-    updated_at = models.DateTimeField(default=timezone.now)
+from .validators import year_validator   # isort: skip
+from users.models import User  # isort: skip
 
 
 class Category(models.Model):
